Limpeza de arquivos desnecessários
E também melhoria na organização dos arquivos das páginas
