--- crint-site/src/dictionary/pt.json
@@ -1,76 +1,0 @@
-{
-    "header" : {
-        "mobilidade" : {
-            "titulo" : "MOBILIDADE",
-            "aluno" : "Alunos(as)",
-            "professor" : "Professores(as)",
-            "servidor" : "Servidores(as)"
-        },
-
-        "estrangeiros" : {
-            "titulo" : "ESTRANGEIROS(AS)",
-            "guias" : "Guias"
-        },
-
-        "informacoes" : {
-            "titulo" : "INFORMAÇÕES",
-            "convenios" : "Convênios",
-            "dia" : "DIA",
-            "pesquisa" : "Pesquisa conduzida"
-        }
-    },
-
-    "homepage" : {
-        "searchbar" : "Digite aqui o que você procura..."
-    },
-
-    "pages" : {
-        "mobilidade" : {
-            "titulo" : "MOBILIDADE",
-        
-            "subpaginas" : {
-                "alunos" : {
-                    "titulo" : "ALUNOS(AS)"
-                },
-                "professor" : {
-                    "titulo" : "PROFESSOR(A)"
-                },
-                "servidor" : {
-                    "titulo" : "SERVIDOR(A)"
-                }
-            }
-        },
-    
-        "estrangeiros" : {
-            "titulo" : "ESTRANGEIROS(AS)",
-        
-            "subpaginas" : {
-                "guias" : {
-                    "titulo" : "GUIAS"
-                }
-            }
-        },
-    
-        "informacoes" : {
-            "titulo" : "INFORMAÇÕES",
-        
-            "subpaginas" : {
-                "convenios" : {
-                    "titulo" : "CONVÊNIOS"
-                },
-                "dia" : {
-                    "titulo" : "DIA"
-                },
-                "pesquisa" : {
-                    "titulo" : "PESQUISA"
-                }
-            }
-        }
-    },
-
-    "footer" : {
-        "endereco" : "Endereço",
-        "redes" : "Redes Sociais",
-        "contato" : "Contato"
-    }
-}
