--- crint-site/src/pages/estrangeiros.scss
@@ -1,7 +1,0 @@
-/* #estrangeiros-root {
-    color: black;
-}
-
-#estrangeiros-root a {
-    color: darkblue;
-} */
