--- crint-site/src/pages/foreigners subpages/guias.scss
@@ -1,7 +1,0 @@
-/* #estrangeiros-root {
-    color: black;
-}
-
-#estrangeiros-root a {
-    color: darkblue;
-} */
