--- crint-site/src/pages/informacoes.scss
@@ -1,7 +1,0 @@
-/* #informacoes-root {
-    color: black;
-}
-
-#informacoes-root a {
-    color: darkblue;
-} */
