--- crint-site/src/pages/information subpages/convenios.scss
@@ -1,1 +1,0 @@
-
