--- crint-site/src/pages/information subpages/dia.scss
@@ -1,1 +1,0 @@
-
