--- crint-site/src/pages/information subpages/pesquisa.scss
@@ -1,1 +1,0 @@
-
