--- crint-site/src/pages/mobilidade.scss
@@ -1,7 +1,0 @@
-/* #mobilidade-root {
-    color: black;
-}
-
-#mobilidade-root a {
-    color: darkblue;
-} */
