--- crint-site/src/pages/mobility subpages/aluno.scss
@@ -1,1 +1,0 @@
-
