--- crint-site/src/pages/mobility subpages/professor.scss
@@ -1,1 +1,0 @@
-
