+++ crint-site/src/paginas/informação subpaginas/pesquisa.tsx
@@ -1,9 +1,7 @@
-import TopicBanner from '../../components/TopicBanner';
-import WIP from '../../components/wip';
+import TopicBanner from '../../componentes/TopicBanner';
+import WIP from '../../componentes/wip';
 import { INFORMACOES_GRADIENTE } from '../../utils/appConstants';
 import { INFORMACOES_BANNER } from '../../utils/appImages';
-import './pesquisa.scss'
-
 
 const Pesquisa = () => {
     return (
