+++ crint-site/src/paginas/mobilidade subpaginas/servidores.tsx
@@ -1,8 +1,7 @@
-import TopicBanner from '../../components/TopicBanner';
-import TopicSection from '../../components/TopicSection';
+import TopicBanner from '../../componentes/TopicBanner';
+import TopicSection from '../../componentes/TopicSection';
 import { BEJE_CLARO, MOBILIDADE_GRADIENTE } from '../../utils/appConstants';
 import { MOBILIDADE_BANNER } from '../../utils/appImages';
-import './servidor.scss'
 
 const MobilidadeServidor = () => (
     <>
