+++ crint-site/src/paginas/mobilidade.tsx
@@ -1,9 +1,8 @@
-import TopicBanner from '../components/TopicBanner';
+import TopicBanner from '../componentes/TopicBanner';
 // import TopicSection from '../components/TopicSection';
-import WIP from '../components/wip';
+import WIP from '../componentes/wip';
 import { MOBILIDADE_GRADIENTE } from '../utils/appConstants';
 import { MOBILIDADE_BANNER } from '../utils/appImages';
-import './mobilidade.scss';
 
 const Mobilidade = () => {
     return (
