+++ crint-site/src/paginas/outros/NotFound.scss
@@ -1,4 +1,4 @@
-@use '../Base.scss' as base;
+@use '../../Base.scss' as base;
 
 #notfound-root {
     margin-top: base.$altura-header;
